+++ react/features/authentication/components/native/WaitForOwnerDialog.js
@@ -6,7 +6,7 @@ import type { Dispatch } from 'redux';
 import { ConfirmDialog } from '../../../base/dialog';
 import { translate } from '../../../base/i18n';
 import { connect } from '../../../base/redux';
-import { cancelWaitForOwner, _openLoginDialog } from '../../actions.native';
+import { openLoginDialog, cancelWaitForOwner } from '../../actions.native';
 
 /**
  * The type of the React {@code Component} props of {@link WaitForOwnerDialog}.
@@ -97,7 +97,7 @@ class WaitForOwnerDialog extends Component<Props> {
      * @returns {void}
      */
     _onLogin() {
-        this.props.dispatch(_openLoginDialog());
+        this.props.dispatch(openLoginDialog());
     }
 }
