+++ react/features/authentication/components/web/WaitForOwnerDialog.js
@@ -6,7 +6,7 @@ import type { Dispatch } from 'redux';
 import { Dialog } from '../../../base/dialog';
 import { translate, translateToHTML } from '../../../base/i18n';
 import { connect } from '../../../base/redux';
-import { openLoginDialog, cancelWaitForOwner } from '../../actions.web';
+import { cancelWaitForOwner } from '../../actions.web';
 
 /**
  * The type of the React {@code Component} props of {@link WaitForOwnerDialog}.
@@ -76,9 +76,9 @@ class WaitForOwnerDialog extends PureComponent<Props> {
      * @returns {void}
      */
     _onIAmHost() {
-        const { dispatch } = this.props;
+        const { onAuthNow } = this.props;
 
-        dispatch(openLoginDialog());
+        onAuthNow && onAuthNow();
     }
 
     /**
@@ -88,7 +88,7 @@ class WaitForOwnerDialog extends PureComponent<Props> {
      */
     render() {
         const {
-            _room,
+            _room: room,
             t
         } = this.props;
 
@@ -102,7 +102,7 @@ class WaitForOwnerDialog extends PureComponent<Props> {
                 <span>
                     {
                         translateToHTML(
-                            t, 'dialog.WaitForHostMsg', { room: _room })
+                            t, 'dialog.WaitForHostMsg', { room })
                     }
                 </span>
             </Dialog>
